Tetra compounds: use drop-down menu
Use a drop-down menu to select which compounds of tetrahedra one is
interested in. This make the user experience less conplicated.

A lot of text is removed as well. The plan is to just give an overview
of all compounds here and have one page with background info, explaining
terms (also using a drop-down) menu.

TODO:
  - create a page with background info
  - move the conjunctive compounds to separate table